requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
faiss-cpu==1.7.3
transformers==4.30.0
torch==2.0.1
python-multipart
